Move sampling priority fully to the context
