+++ tests/test_sampler.py
@@ -4,10 +4,11 @@ import unittest
 import random
 
 from ddtrace.span import Span
-from ddtrace.sampler import RateSampler, AllSampler, RateByServiceSampler, SAMPLE_RATE_METRIC_KEY, _key, _default_key
+from ddtrace.sampler import RateSampler, AllSampler, RateByServiceSampler, _key, _default_key
 from ddtrace.compat import iteritems
 from tests.test_tracer import get_dummy_tracer
 from .util import patch_time
+from ddtrace.constants import SAMPLING_PRIORITY_KEY, SAMPLE_RATE_METRIC_KEY
 
 
 class RateSamplerTest(unittest.TestCase):
@@ -91,11 +92,11 @@ class RateByServiceSamplerTest(unittest.TestCase):
             samples = writer.pop()
             samples_with_high_priority = 0
             for sample in samples:
-                if sample._sampling_priority:
-                    if sample._sampling_priority > 0:
+                if sample.get_metric(SAMPLING_PRIORITY_KEY) is not None:
+                    if sample.get_metric(SAMPLING_PRIORITY_KEY) > 0:
                         samples_with_high_priority += 1
                 else:
-                    assert 0 == sample._sampling_priority, "when priority sampling is on, priority should be 0 when trace is to be dropped"
+                    assert 0 == sample.get_metric(SAMPLING_PRIORITY_KEY), "when priority sampling is on, priority should be 0 when trace is to be dropped"
 
             # We must have at least 1 sample, check that it has its sample rate properly assigned
             assert samples[0].get_metric(SAMPLE_RATE_METRIC_KEY) is None
